+++ src/plugins/calculate-next-release.ts
@@ -5,9 +5,9 @@ import semver from 'semver';
 const { SemVer } = semver;
 
 import { BoneheadedError, isReleaseProps, ReleaseHeading, ReleaseProps } from '../types.js';
-import { PrepareReleaseOptions } from '../options.js';
+import { BumpOptions } from '../options.js';
 
-const attacher: Plugin = function (options: PrepareReleaseOptions) {
+const attacher: Plugin = function (options: BumpOptions) {
   // eslint-disable-next-line @typescript-eslint/unbound-method
   const processorData = this.data;
   return transformer;
@@ -25,12 +25,7 @@ const attacher: Plugin = function (options: PrepareReleaseOptions) {
 
     const latestVersion = latestRelease ? (<ReleaseProps>latestRelease.release).version : new SemVer('0.0.0');
 
-    file.data['nextReleaseVersion'] = semver.inc(
-      latestVersion.format(),
-      options.releaseType,
-      undefined,
-      options.prereleaseIdentifier
-    );
+    file.data['nextReleaseVersion'] = semver.inc(latestVersion.format(), options.version, undefined, options.preid);
   }
 };